Move cookie consent definitions from @app to @persistence
It fixes a cycle between @app and @persistence.
